requirements.txt
fastapi
uvicorn
google-adk
pydantic
python-dotenv
